requirements.txt
# HTTP Client
requests>=2.31.0

# Environment Variables
python-dotenv>=1.0.0

# Data Processing
pandas>=2.0.0

# Type Hints & Validation
pydantic>=2.0.0
